NEVE_python/requirements.txt
mlagents-envs==0.16.1
PyYAML==6.0
pyserial
Gooey
nuitka
zstandard
protobuf==3.20